requirements.txt
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
idna==3.4
paramiko==3.3.1
psutil==5.9.5
pycparser==2.21
PyNaCl==1.5.0
pyTelegramBotAPI==4.13.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
telebot==0.0.5
urllib3==2.0.4
